Throws error if attempting to delete a non-empty bucket 🎳
See https://docs.aws.amazon.com/AmazonS3/latest/API/API_DeleteBucket.html

#233

+++ src/main/java/ninja/S3Dispatcher.java
@@ -92,6 +92,7 @@ public class S3Dispatcher implements WebDispatcher {
     private static final String RESPONSE_BUCKET = "Bucket";
     private static final String ERROR_MULTIPART_UPLOAD_DOES_NOT_EXIST = "Multipart Upload does not exist";
     private static final String ERROR_BUCKET_DOES_NOT_EXIST = "Bucket does not exist";
+    private static final String ERROR_BUCKET_IS_NOT_EMPTY = "Bucket is not empty";
     private static final String ERROR_FILE_SYSTEM = "Problems with file system access";
     private static final String PATH_DELIMITER = "/";
 
@@ -467,6 +468,15 @@ public class S3Dispatcher implements WebDispatcher {
             if (!bucket.exists()) {
                 signalObjectError(webContext, bucketName, null, S3ErrorCode.NoSuchBucket, ERROR_BUCKET_DOES_NOT_EXIST);
             } else {
+                if (bucket.countObjects("") > 0) {
+                    signalObjectError(webContext,
+                                      bucketName,
+                                      null,
+                                      S3ErrorCode.BucketNotEmpty,
+                                      ERROR_BUCKET_IS_NOT_EMPTY);
+                    return;
+                }
+
                 if (!bucket.delete()) {
                     signalObjectError(webContext, bucketName, null, S3ErrorCode.InternalError, ERROR_FILE_SYSTEM);
                     return;
